idp-oidc/requirements.txt
fastapi
uvicorn
cryptography
PyJWT[crypto]